could work

+++ include/common.h
@@ -7,6 +7,6 @@
 
 // 根据平台判断，目前支持 ARM-Linux
 typedef uint64_t addr_t;
-typedef uint32_t instr_t;
+typedef uint64_t instr_t;
 
 #endif
